+++ packages/frontend/.eslintrc.js
@@ -24,10 +24,9 @@ module.exports = {
         'airbnb-typescript',
         'prettier',
         'plugin:jest-dom/recommended',
-        'plugin:testing-library/react'
+        'plugin:testing-library/react',
     ],
     plugins: [
-        '@typescript-eslint',
         'css-modules',
         'import',
         'json',
@@ -36,7 +35,7 @@ module.exports = {
         'react-hooks',
         'react',
         'jest-dom',
-        'testing-library'
+        'testing-library',
     ],
 
     settings: {
@@ -94,15 +93,15 @@ module.exports = {
             'error',
             {
                 prefer: 'type-imports',
-                fixStyle: 'inline-type-imports'
-            }
+                fixStyle: 'inline-type-imports',
+            },
         ],
         '@typescript-eslint/no-unused-vars': [
             'error',
             {
                 argsIgnorePattern: '^_',
                 varsIgnorePattern: '^_',
-                ignoreRestSiblings: true
+                ignoreRestSiblings: true,
             },
         ],
     },